windows/requirements.txt
flask==1.1.2
gunicorn==19.9.0
requests==2.23.0
flask_restful==0.3.8
flask_cors==3.0.8
pymysql==0.9.3
flask-compress==1.8.0
flask-caching==1.9.0
psutil==5.8.0
timeloop==1.0.2
pywin32==300
waitress==2.0.0
pyautogui==0.9.52
werkzeug==1.0.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==5.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1